Separate the position from Game
