+++ crates/grob_core/src/primitives.rs
@@ -7,5 +7,5 @@ mod square;
 pub use bitboard::BitBoard;
 pub use board::Board;
 pub use castling_rights::CastlingRights;
-pub use pieces::{Color, Piece, Promotion};
+pub use pieces::{piece_from_fen, Color, Piece, Promotion};
 pub use square::{File, NegDiag, PosDiag, Rank, Square};
